build: Enable standalone build (#11)
This PR adds the capability to build GPCC standalone inclusive
building a unittest executable.
If GPCC is included by a top-level project, then there is no change
in behaviour. If GPCC is build stand-alone, then a unittest executable
is build if the target environment is "unittest".

Scripts for triggering all build steps are provided.
A .vscode folder with settings and tasks is added.
Optional auto-detection of configuration is added.

---------

Co-authored-by: Wishmaster-de <74838617+Wishmaster-de@users.noreply.github.com>