fix(osal)!: Clock granularity in convar timed wait (#12)
With this commit, ConditionVariable::TimeLimitedWait() will
not add any compensation for clock granularity to the timeout
value across all platforms. The timeout is an absolute point in
time. If the user requires a minimum time span, then he/she
shall add the clock's granularity to ensure a minimum time span.