build: Add support for GCC UBSan (#17)
This commit adds support for building the unittest executable
with GCC's Undefined Behaviour Sanitizer.

The build-configurations for unittests were revised and cleaned
up. There are three flavors of unittests now:
build_unittest:
For everyday use in develoment.
build_unittest-buildsrv:
For the build server. Applies UBSan and memcheck.
build_unittest-notfc:
Builds unittest without presence of TFC. This enables a few tests
for the OSAL that require absence of TFC. This is used on rare
occasions and only locally.