fix: Use snprintf() instead of sniprintf() (#34)
sniprintf() (and friends) are a newlib-specific extension to the
standard C-library, which support integer types only. The idea is to
use only the i-variants if floating point data types are not used in a
project and thus safe a few kB of code.

Before this commit, GPCC used sniprintf() in cases where only integer
types are involved and where newlib is the C-library. To detect if
newlib is the C-library, presence of the preprocessor define
_NEWLIB_VERSION was tested.

However, on Zephyr _NEWLIB_VERSION seems to be defined, but the
functions are not available resulting in compiler/linker errors.

However, in large projects (>150kB of code) there is no significant
benefit from using sniprintf() (and friends) because:
- floating point types may simply be required
- there is plenty of flash and the few saved kB of code
are not worth compared to the total size of the project

This commit removes any usage of sniprintf() (and friends)
from GPCC and replaces it with snprintf() and friends.
Closes #31