feat: CMake generates doxygen documentation (#36)
This commit adds a build target `gpcc_doxygen` for building the doxygen
documentation by the build system generated by CMake. The shell scripts
and doxygen files for different OS are still in place to allow for quick
and easy generation of the documentation without using CMake.

A CMake variable `GPCC_DOXYGEN_OUT` is exported to the parent project.

---------

Co-authored-by: Wishmaster-de <74838617+Wishmaster-de@users.noreply.github.com>